% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ahr.R
\name{print.ahr}
\alias{print.ahr}
\title{print.ahr}
\usage{
\method{print}{ahr}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class '"ahr"'.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print ahr object
}

