% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmean.R
\name{print.rmd}
\alias{print.rmd}
\title{print.rmd}
\usage{
\method{print}{rmd}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class '"rmd"'.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print rmd object
}

