% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantiles.R
\name{print.survQuantile}
\alias{print.survQuantile}
\title{print.wkmQuantile}
\usage{
\method{print}{survQuantile}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class '"survQuantile"'.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print survQuantile object
}

